requirements.txt
autopep8==1.5.6
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pycodestyle==2.7.0
pyEDFlib==0.1.20
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.2
simpleaudio==1.0.4
six==1.15.0
toml==0.10.2
